pyproject.toml
[project]
name = "not-cat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.1.3",
    "onnx>=1.17.0",
    "pillow>=11.0.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.0",
    "ultralytics>=8.3.23",
]
